Standardize the use of job object as a runner attribute
The only exception to this is the `run_suite()` function where
the job is first passed as an argument. The only affected module
by this change is the runner module.

The previous use of the job object was inconsistent.

Some future todos will include simpler node environment setting
possibly dropping the entire `set_environment` method.