Enable per-slot vm operations like boot/download/upload/shutdown
These tools have to be performed once per worker environment without
any regard on other workers. We can thus utilize a simple always-true
run decision policy. A current limitation is that the workers would
wait for each other to traverse the setup node (pseudotest) but this
limitation will be lifted once we allow node reentrancy based on the
same run decision policy (e.g. for parallel test retries).

Also adapt the sample test development tool since it is based on the
same implementation as all of the above.