Add support for multi-variant external states
This required some exploratory and context refreshing isolation
tests for the test object parsing capabilities of the test loader.
In particular, these tests were used to identify what is the best
way to parametrically (for state backend access) track all the
variants of the components of a net composite and by extension a
test node. It also involved some docstring clarifications regarding
the test object parsing interface.

For all current state backend logging we will still rely on just
the vm suffix instead of the whole variant name since the latter
must be inferrable by means of the currently run tests.