Fix incorrectly applied vm object restrictions via test node params
Previously it was possible that a restriction of one suffix variant
for a vm like only_vm1 leaks into choices of vm2 variants by matching
the wrong part of the net variant name and thus producing incorrect
node variants as a result. The workaround for this was a rather
repetitive

only_vm1 = vm1.qemu_kvm_centos

in order to match the right net and vm1 variant. Now the original
configuration works properly

only_vm1 = qemu_kvm_centos

A second undetected problem of the previous implementation was that
restriction of a given variant might also apply to another if the
two are parsed separately. The reason was that a more restrictive
initial test set would result in fewer (but not zero) vms and if
these are then reused by a second test set, no new nets and vms
will be parsed even if such are still needed.

Solve both problems by completely rewriting the nets get and parse
functionality with isolation tests that detect the above problems
and improve the overall readability of the complex composite net
retrieval for each node.