Convert the previous slot ids to worker runtime strings
The parsed workers are now utilized during traversal and in all
node related operations, providing the original slot runtime string
whenever necessary. Both parsing and running of workers had to be
refactored for proper reusability from manual tools, the former in
order to provide runtime slot workers for separately parsed objects
and nodes and the latter to allow for job-bound spawner to be set
together with startup and other worker preparatory steps.

This also fixes an integration test fallout due to incorrect worker
id usage for container set up.