Split the old object parsing into suffix and component parsing
We need a temporary method for net from object strings in order to
provide the old behavior to callers which implicitly parsed components
for a default net (net1).

The additional unflatten flag overlaps with some of the composite
object parsing but helps us work around newly discovered Cartesian
config limitations when parsing composite nets by reusing the older
implementation of doing so.