Migrate all parsing functionality to the Cartesian graph class
All changes in this migration other than copying the same code are
only about conversion to static methods with the exception of a few
of the previous loader methods still using graphs that now use the
actual graph instance.

The test loader has now been turned into a simple boundary module
dealing only with the Avocado interface while the test runner now
also accepts a test suite of runnables obtained from flat test nodes.
Thanks to this we can also get rid of an ancient hack due to the
fact that we only work with Cartesian graphs while Avocado assumes
a simple list of tests to be passed around to the runner.

Also use the opportunity to make the test loading tool use standard
job-provided log path for visualizing the test graph.