Choose net suffix from flat object when parsing objects for object
This is the first step to providing either async graph parsing
for each test worker or better yet on demand parsing during
async traversal of partial graphs.

Thus finish type annotation and improved argument order in the
parsing functionality API of the Cartesian graph class with the
exception of the object tree parsing method.