Provide more natural support for flat nodes via the node class
Constructing test nodes with an obligatory test object was meant
only for instantiating full instead of flat test nodes. The test
node class also needs a few updates in order to no longer assume
a node instance is always constructed with a test net, some of
which have the extra benefit of explicit type calls.

The flat node construction is now similar to the construction of
flat test objects with properly generated parameters and some
shared considerations when it comes to performance and the use
of generators.