Migrate all traversal functionality to the Cartesian graph class
All changes in this migration other than copying the same code are
about making the node traversal and reversal methods public and
formal part of the API with more up-to-date renaming for all methods.

The test runner has been reduced to test running functionality and
test status tracking and is provided to the constructed graph via
the worker preparation stage (and is thus swappable with different
implementations).