Formalize node, net, and vm parameter mixing
In particular, the vm variant restrictions of test nodes and flat
test nets should not reach the vm-specific parameters (will result
in "only" parameter key and is not valid within the vm scope to
begin with) and they should both be used to filter all reused
and parsed vms. Due to a Cartesian parser limitation, only flat
nets could contain vm restrictions and full ones should be free
of such so that we could parse nodes from them without overwriting
the node vm restrictions. They use alternative parameter for the
restriction which is also taken into account if objects are parsed
or reused for a full net instead of a flat one.

The flat net is meant to contain the objects overwrite configs
where default vm restrictions are first defined and must thus be
used for such defaults to be properly applied in all use cases.
This requires the default flat net that we currently parse for
backwards compatibility to be less of a shallow solution and a
better imitation of a typical worker net.

Also make sure overwrite params are defined in a compact manner
whenever necessary and improve such object restrictions in the
object node parsing method.