Make the test nodes runnable with default parsing on demand
This means we can now properly use the test loader resolution of
flat nodes as runnables to be fully parsed while being traversed.

One requirement of test nodes becoming runnable is to improve the
name of the recipe for parsing them (and thus avoid collision with
the actual config attribute of a runnable which is more appropriately
named). A secondary such is to update all VT parameters (used by the
environment local VT runner) when regenerating parameters, dropping
unnecessary complexity wherever possible.