Add isolation tests of all newly accessible parsing up to cloning
This comes with a few minor improvements on parent parsing verbosity,
shared pool fix on slotless mock runs, and anything that was necessary
to move from single as easy and as default as possible test for each
newly accessible parsing method to multiple isolation tests for it
that now include more complex and more general cases.

It even helped us find a new problem with branch parsing that involves
cross object leakage and should be fixed next to re-enable one of the
isolation tests.