Add setup locations in a parsing-on-demand compatible manner
The previous implementation assumed all cleanup nodes of a
node already exist which is no longer the case.

A better way to implement the same is to use worker instances
in order to get rid of some of this configuration boilerplate.

Also use correct remote swarm setup port expectation since "22" is
contained in the correct one to assert "222".