Use a prefix tree with a multi-entry hashmap
The prefix tree together with the hashmap serve as auxiliary
data structure to speed up lookup and retrieval of test nodes,
i.e. an index.

Something to still carefully consider is if the `nodes` attribute
of the graph should remain public or we should instead provide any
removal, replacement, and other functionality to the prefix tree
and only access the nodes attribute via few allowed methods.