Skip redundant state scanning based on global scope information
The pool scope was already utilized globally in order to evaluate
if a given test node is eagerly or fully finished and can be used
in a similar manner (fixing the most granular case of each worker
being separately accountable). This then reduces the number of
performed state scans (involving running remote control files and
a lot of other overhead that slows down the overall traversal) to
just one per swarm for isolated swarmst.

It also finally formalizes the previously oversimplified failed
test rerunning by a mere state scan where instead of retrying a
test node in an extra worker that might or might not traverse it,
it is better to error out with missing state as in all other cases.