Simplify the previous job tests handling to validate test duration
Whether a test will be skipped based on status from a previous job
will now be decided within the run decision policy by considering
a replay status that we should not deviate from (skipping tests
with more than a replay status). In this way a test that passes in
a previous job will be skipped if we are replaying a FAIL or ERROR
test status. For instance, it doesn't make sense to replay a failed
test that then passed. The previous behavior should also be fully
preserved as we used to skip tests with a different than the replayed
status and finding even one such would result in the test being
added to the list of skipped ones.

Parsing results from previous jobs is now simpler and allows for
some extra validation of the test duration whereby the status will
be changed to WARN if the test takes much longer than a reference
result.

Use the opportunity to better place the should_rerun and should_scan
evaluation methods as components of the default run decision. Proper
testing and contracts for the should_replay decisions which are now
more testable away from the project boundary, should be performed
via the retry isolation tests merging the two rerun decisions.