Provide some indexing and recipe for the test objects too
This is meant to complete the use of indices as well as recipe
attributes for both nodes and objects. So far it was done only
for nodes not to interfere with the overall intent of all previous
changes.

With both test nodes and test objects now indexed we will take the
opportunity to protect read-only arguments and accordingly improve
the string representation of a node/object recipe.