Bridge worker-related test nodes to recover previous behavior
In particular, share occupation status, workers, and test results
across bridged nodes and reuse registered picked and dropped nodes
across bridged nodes as edge registers.

The picked-by criterion is inverted in order to be properly devoted
to its own function which is to detect workers that have started
traversing a test node but to avoid picking the same occupied setup
from a different path that wasn't traversed before. Sorting out the
parents or children to pick will also have to prioritize flat nodes
(among which most importantly unrolled nodes by traversed workers)
in order to avoid waiting for too long or even getting stuck and
unrolling all nodes as fast as possible.

The picked-by registers are already a slight improvement of the
earlier divergent paths implementation as can be seen in a few
isolation tests that now have improved order of run test nodes.
Most importantly though, they can be enhanced further via counters
for graph ergodicity which will be our next step.

Cloning had to be improved to inject a clone identifier rather than
just append it for a uniform treatment of the newly introduced node
bridged form to properly identify a "worker generalized test node".
The order of parsing of nets and vms and the way this interplays
with double-join operation severely limits the ease of use as well
as implementation of this bridged form. However, we will have to
live with it until the Cartesian config API and capabilities are
improved enough to allow the original intent of the bridged form.

Since traversal is now restricted to just worker-specific nodes
the manual tools are no longer compatible and will have to be
disabled (as test contracts) until their parallelism support gets
updated with multi-graph support.