Simplify the finished workers attribute to a single finished worker
Now that nearly all nodes can only be traversed by a single worker
they were parsed for, we no longer need yet another set to store
the workers that finished (so far equivalent to traversed) a particular
node. The only exception to this is flat nodes which are already
handled separately in many situations with a few more included now.

The picked-by node registry is now the only one used for divergent
paths selection and the dropped node registry is the only one used
for setup/cleanup readiness and thus reversal considerations. The
"workers" registry is thus now free to be replaced with something
more appropriate for its remaining use cases like "finished_worker".

To better compare to the current worker, we rename the latter to
"started_worker" and will further use it to allow reentrancy of
otherwise occupied nodes.