Refactor started/finished eagerness via a threshold
Define started/finished methods with threshold of 1 to represent
maximum eagerness (e.g. previous is_eagerly_finished) and threshold
of -1 to represent maximum patience (e.g. previous is_fully_finished).
An alternative is_started method to the unified is_finished one is
then used to evaluate if a node is eagerly or fully occupied.

Occupation status is now relative to a worker in order to allow for
swarm and other scope dependent evaluation with max_concurrent_tries
being used as the threshold (default is maximally eager occupation
as before).

Finally, whether a node should scan states is just a negated version
of whether the node was eagerly finished (refactored) and the graph
update tool now supports pool scopes for free.