Restore multi-graph traversal for various tools parsing on their own
While traversing separate nodes per worker still allowed for workers
to traverse nodes parsed for other workers and relied on simple skip
as part of the run/clean decision policy, bridging worker-related
nodes now requires that a node can only be traversed by the worker
it has been parsed for unless it is flat. This means that tools that
used to traverse the same node by various workers and haven't taken
care of this requirement now have to be updated for full multi-graph
traversal compatibility. This in turns simplifies the custom run
policy for some of them as a beneficial side effect.

Let us also use the new parsing API to allow multi-product use of
these tools and handling for incompatible workers where applicable.

The additional permanent vm creation tool was also refreshed with
new API calls and better integrated with an internal update call
(with the update cache tool to look into next).