Restore multi-graph traversal for the main update-cache tool
This needed extension of the graph flagging capabilities with
per-worker node children run/clean policies. Multi-graphs have
separate nodes for each worker and as such allow for worker
customizable run/clean policy that we should take advantage of.

Dropping the old style print verbosity in the update-cache tool also
led to a small improvement of flexibility in the no longer used old
simultaneous parsing of objects and nodes.

The use of from- and to- state parameters is also already simpler
and mimics the typical use of limit-inclusive ranges in practice.