Refactor parse-and-get behavior for flat and parent node expansion
In particular, reuse already parsed nodes (e.g. as dependencies)
when unrolling flat nodes.

The previous flat node expansion was relying on the fact that flat
nodes with reusable composite children that are already cached in
the graph don't have to be unrolled and will already be considered
when traversing these children. However, this has a problem with
multi-graphs because the reusable dependencies might be explored
and parsed for some workers but not for others (previously all
workers traversed the same composite children of the same flat node).
It also fails to consider that only some of the composite children
might be parsed and reusable while others might only remain accessible
via the unrolled flat node.

Let us simplify this behavior by making sure flat nodes are always
unrolled and properly connected with reused and new composite nodes
and test this on three different levels - at the refactored method
level, the branch parsing level, and a graph motif where a test node
of "client_clicked" variant is parsed both as a parent/dependency
and as an unrolled flat node by two respective workers.