Make the setup/cleanup readiness independent for each worker
While the previous improvement of always unrolling flat nodes for
each worker handles many cases of partial exploration, the setup
and cleanup readiness checks still contain an implicit assumption
that all composite children of a node (edge registers via bridged
form) are already parsed and exist at the time of the check. This
assumption can then be violated by any external or future call and
should rather be handled within the checks.

There are also already cases where flat nodes can be unrolled with
reusable nodes that can be cloned or otherwise modified which are
too complicated to handle in any other way. The simplest option
we thus have is to allow for each worker to independently consider
setup/cleanup readiness as if it was the only subgraph we have.