Add a uniqueness flag to parse or get unique flat objects or nodes
This is mostly a syntactic API sugar meant to simplify the use of
the Cartesian graph API providing both validation and unwrapping
of singular test objects and nodes.

Also use the opportunity to rename "get_*_by()" to just "get_*()".
All simplified isolation tests are examples of the simpler API usage
that follows and parsing a unique shared root will already use
updated calls.