Add support for worker-specific incompatible nodes
Cases of never expanded flat nodes must now be handled separately
for each worker since some workers may only support certain vms
and not others thus being unable to expand test nodes with such
incompatible vms. Overwriting a method with a lambda function
would always be done by the first worker to try to expand the
flat node and skipped by (thus enforced on) all others. It also
made such cases hard to observe from the outside and thus test
and assert. Therefore simply register incompatible workers to
allow separate unrolling per worker as well as properly test
this behavior (an example of test driven design).

The parsing on demand restriction meant to test this was incorrectly
applied and thus futile from the introduction of the test but would
have kept the test failing for most of the multi-graph changes so
let us recover it first with the current change making unexpandable
nodes compatible with the multi-graph. Most importantly, test these
incompatible nodes explicitly now (if we had done so in the past the
futility in the original test would have been caught earlier).

Also allow cleanup of flat nodes to reduce postponing overhead
since the only flat node we may add adhoc children to is the
shared root.