Apply default vm restrictions in graphs parsed on demand
Separate any node and object parameters from object restrictions
provided as restrictions parameters. Parameters of the form
`only_suffix` that were previously used to identify object cache
(incl. nets) for a test node support overwriting which is still
desirable for only/no filters but only in a few specific cases.
Having a specialized restrictions attribute and restriction-free
parameters allows for safe overwriting at various stages and
allows for the restrictions to either turn into further multi-
line restrictions or to overwrite (if still used as a parameter).
This was partly inspired by the command line parser which does
a trick like this to extract vm restrictions from command line
parameters and even provides a "default restriction" via a
`default_only_*` parameter in the configs.

Add object/node retrieval by a multi-line restriction that is
capable enough to handle a separate "only/no" filter per line
and an OR operator "," via regex translation (AND operator ".."
can be emulated by adding multiple "only" filters as multiple
lines). The retrieval provides optional objects/nodes cache via
the graph instance.

Parsing a full composite object (mostly net) from other objects
can now be simplified as we don't need an `object_only_*` parameter
to work around overwriting it conditionally or otherwise. Another
simplified but also enhanced in terms of functionality cases are
parsing a node from an object, parsing objects for a node and object,
and most importantly parsing branches for a node and object, all of
which make use of the restriction-based object/node retrieval to
check for compatibility.

Previous object strings have to be formalized as restrictions with
a more standard shared net that doesn't have to be mocked at runtime
and simply converts the object restrictions to supported such. This
then improves the parsing of objects and nodes avoiding conversion
of the previous strings into restriction parameters and other cases
where we no longer have to filter out such restriction parameters.

Providing a dictionary of restrictions indexed by the object suffixes
allows us to also drop the redundant final restriction of test nodes.

Finally, flat objects are now officially considered as well.

Some todos still had to be left in the changes but will be handled
next in order to complete this subbranch.