Define additional serial and remote nets via cluster products
This allows the remote work tracing test to use default parameters
similarly to all other work tracing tests but with cluster-aware
get location suffixes (not necessarily the net suffixes). It also
makes serial runs use explicit worker ID (net0) in order to be more
compatible with the current formatting and any node/object retrieval
relying on the worker ID. The net0 is intentionally excluded from
the default available nets since it is used only for legacy serial
runs.

The multiple cluster variants contain localhost which is used for
both serial runs (process isolation) and local LXC runs (container
isolation). A more appropriate named "nets_id" is now used instead
of the previous LXC container assuming "cid". One limitation that
still remains though is the need for unique net variants which are
still widely used as worker identifiers.

The serial net can also be used for test listing which is serial in
nature and doesn't have to be performed for all nets by default.