Resolve parameter collision using separate dependency object id
In some cases of supported vms restriction by workers the object ID
and suffix parameter meant to be parsed in the sense of "dependency
for" interfere with the identically named parameters for object ID
and suffix meant as "ID-s and suffixes of object components". Since
the latter are used to identify the image and state cache for the
specific vm variant, they might be wrongly overwritten by the former
dependency parameters, making the vm cache folder inaccessible. Let
us clean up the mixed interpretation and overwriting of these params
before cleaning up the mixes with "only_suffix" restrictions.