Allow reuse of flat nodes during composite node parsing
This is an important interface improvement which prevents mixing
of vm restrictions for a flat node and improves performance by
reducing the need of multiple parsing steps for the same flat nodes
and parsing of polynomial number of composite nodes and nets before
additionally restricting them instead of doing all of this earlier.

Since the parse_and_get version has to be doubled just to make this
available let's make it more argument-flexible instead.

The flat nodes can now reuse previous restriction logic and will
avoid parsing full nodes for many unnecessary possibilities with
filtering outsourced to the object selection and parsing method.