Make the workers globally accessible via the run slots
Since the previous run slot overwriting has been well established
and we need access to all workers in multiple situations, let's
take the best of both worlds and provide the workers globally for
various swarm-related decisions.

The swarm filtering pool scope is not fully tested since it is
an area to redesign with swarm development further steps.