Choose more odd number nets to include OR-ed vm restrictions
This allows us to test cases like "only A, B" and "no C, D" for both
local and remote containers: how they are parsed, overwritten, and
where they take effect. Due to a current limitation in the Cartesian
config these (and any) net parameters cannot be overwritten and only
edited in the config file. However, this is reasonable as these are
only meant to be rarely changed supported vms and the command line
restrictions of vm variants would rather rely on overwriting (by
further restricting) flat test node parameters. This most standard
approach is now also used in all relevant tests and refactored in
its own reusable method similarly to the reusable graph traversal.
We thus only had to reduce the number of workers of the parsing on
demand test to exclude net3 and its overwriting limitation as the
test emulates a different supported vm variant for its own validation.

There are some problems with the prefix priority and ordering of vm
variants that this standardization of selected object restrictions
found and will be solved once this subbranch is merged.