Restore fixed order of component restrictions
Instead of using an ordered dictionary, we can simply provide safe
ordering for the suffix variants of all object types (vms and nets)
that is based on explicit user-predefined choice and validate that
the base "vms" and "nets" parameters contain and stay in full sync
with the object suffix variants that are actually used.

This in turns requires explicit listing of all possible net suffixes
(serial and cluster products included) in order to validate each.