Reset traverse path instead of storing discontinuous parts
A worker should start a new traverse path from the shared root node
each time it encounters occupied node or postpones cleanup reversal
due to more unrolled flat nodes. It will perform additional asyncio
sleep in the first case to allow for occupied nodes to free up but
not in the second case to keep asyncio interruptions and order simpler.
This is possible because of the graph ergodicity and the guarantee
that a worker would ultimately return to the same node after long
enough time.

The order of tests also has to be adapted in some bloated work tracing
cases and described again.