Simplify setup locations pull by separating net configuration
Use the test results to collect setup locations instead of manually
propagating runtime parameters through nodes together with the
traversing workers. Setup location parameters still have to be
updated on each traversal step for the same node and suffixed by
the nets they refer to.

We use worker ID-s to identify each location environment and all of
its configuration, be it gateway, host, port, or else. The idea is
to still allow for multiple paths per such environment as before
but replace the excessive (and not even fully compatible) location
path suffixes with the simpler worker net suffixes.

Also use the opportunity to update the new order of traversal for
traced work which we couldn't do until these tests pass.