Drop runtime string information from state scan/sync
Unfortunately for now these methods will have to make the implicit
assumptions that the `started_worker` attribute is already set but
this is still the case and it was also assumed in the past via the
expectation of `nets_host` and `nets_gateway` parameters.

Further refactoring of the external state diffchains is needed but
goes beyond the scope of this branch.