Reuse clones of all generations if these are already parsed
As cloning was barely tested in isolation before there was generally
no guarantee on full usability until we could give such now. Reusing
the clones is now also tested in isolation and done by returning the
correct already available dependency, separately found clones via
subvariant parsing, or full replacement of a clone source by its
clones in cache.

Reusing the clones was not possible with the in-place changes of
the clone source which was mandated by the fact the clone source
is already cached in the graph and fully connected to its setup
and cleanup nodes. The new approach will reduce such side-effects
by returning the clone source together with all its explicit clones
allowing the replacement of the latter with already parsed cache.
While this slightly modifies the interface of branch parsing, it is
also better to return instead of implicitly remove the clone source.

Since the graph and node indexing via prefix tree are additive
data structures only removing the clone source is not a viable
option. This is further affirmed by the fact that it will be
parsed again from flat nodes or in other cases and would have to
be reused multiple times. Let us instead prevent the clone source
from running/rerunning and cleaning/syncing and give it a special
0-prepended prefix.

Bridging had the extra effect of preventing double parsed clone
nodes from rerunning and thus easy detection but in this way it
was hiding unexpected behavior where parallel reruns could pick up
different otherwise identical nodes.

The previous assertions on non-yet-added setup and cleanup node
relations are no longer valid since the full node child of a flat
node could have its dependency parsing done again. While this also
used to be the case before, it was hidden by an implicit behavior
of the dependency parsing method which was hiding and not returning
the reused partially parsed dependency (excessive coupling) and now
explicitly returns such parents too.