Refresh all nontrivial node getters
Most importantly, return the dependency test node that will be
needed for reusable cloning. Most generally, return test objects
and nodes for all such getters and not just for the terminal object
getter. Getting a list of objects/nodes is also something we will
have to do for clone sources when restoring the deep cloning next.

These are nontrivial getters because their behavior can be tweaked
via passed arguments.