Protect node relations from each node to all other nodes
Cloned and bridged nodes as well as setup and cleanup nodes are
all fully private now with properties returning a read-only version
of the otherwise mutable list and dictionary upon retrieval and can
only be modified via the corresponding symmetric methods for clone
sourcing, descending, and bridging.

Use the opportunity to provide public access and thus isolation
test for the previously static and private branch cloning now that
cloning is properly tested and by extension properly supported.