Restrict the fully started/finished node checks to involved workers
This simplifies the complexity from having to consider all available
workers and register incompatible ones at the right time to avoid
an incompatible worker being the last one to clean up the node and
other difficult circumstances where we cannot guarantee the right
cleanup conditions. We have to allow parsing to be skipped in order
to allow the current set of workers the node was picked up by to be
sufficient for cleanup. The default cleanup policy should then be
update to block cleanup not only if the node is finished for all
involved workers but also if it is cleanup ready for these.

The three main cleanup consideration stages are in this order:
  - `is_cleanup_ready` (for cleanup nodes and traversal)
  - `should_clean` (current and reversible nodes)
  - `is_finished` (scope and finished markers)
where the default cleaning decision via `should_clean` will now
also make use of `is_cleanup_ready` together with `is_finished`
to evaluate on all involved workers.

Also perform a few cleanup improvements where the worker variable
name was colliding with a method argument and flat nodes could use
some speedup as they are not reversible. The unknown status (still
running) tests should also be considered per involved worker just
in case.

In short, last one to close the door is considered from all that have
opened the same door so far rather than all possible visitors.