Do not register nets with reused composite nodes as incompatible
The newly introduced reusability of cloned nodes and speedup for
unrolled flat nodes could end up with the path parsing method not
returning any newly parsed siblings which would be wrongly detected
as a flat node that does not compose with a given worker's net. Let
us fix this and improve incompatible worker nets detection by better
localizing the check, warning, and registration of such cases to the
unrolling of a flat node during branch parsing.

One sacrifice this needed was the use of long suffixes of the worker
nets rather than the complete workers themselves due to the way such
information is distributed but this does not limit the logic of how
the incompatible workers are used.