Add multi-product support for the update-cache tool
This tool is capable of updating on different products of the same
vm at the same time and such use cases make sense. It does not make
sense for cases like booting multiple products of the same vm or
retrieving states for multiple products of the same vm, i.e. the
previously restored tools.