Avoid diffchain loops if the set state already exists
Overwriting a sourced state backend needs more sophisticated handling
than that of sourced state root backends where each state is rather
derived from the same starting point even if they are in a linear
succession. This means that we can no longer apply the logic of
simply removing previous states before recreating them and should
instead delegate this and further decision making to each backend
itself (for any of nontrivial/nonroot states).

While the ramfile backend mostly mimics the logic of the qcow2ext
backend, the latter concentrates some non-imperative enhancements:

1) Avoid circular references if getting and setting the same state
by detecting if the current overlay will be saved as its own already
existing backing file and commit instead.

2) Also allow safe forward overwrites to still support update-cache and
other manual tools.

3) Prevent any further nontrivial overwriting process which needs
more manual action and thinking (and that the user knows what they
are doing).

An undesirable side effect of this is that we no longer retain the
strictly imperative boundary (functional core, imperative boundary)
but we can only frame this in a todo since the state backends need
deeper changes and improved conceptualization to recover it.

The isolation tests for the state setup need a bit more enhanced
class mocking in order to pass the newly introduced subclass check.