Standardize extra and early logging details for the plugin
In particular, provide namespaced logging for both parsers
but use the default avocado plugin approach for manu/auto.

Get rid of extra logging levels that we can easily mimic and
thus accomplish the same effect from using our own log viewer
filtering.

Provide a stack trace for errors within the manu tool but be
explicit about early logging flags needed to see it. This is
still not breaking the experience for the user but provides
useful debugging logs when needed.