Add only setup-access relevant parameters when pulling locations
Update test node parameters only with "nets_" prefixed parameters
which are needed for accessing the setup at the specified location.
Remaining parameters are not relevant and worse yet could lead to
unpredictable vmnet parameters (incl. name) and thus test node
parameters updated using the respective vmnet.

If we have a worker provide vm/image state setup for a given test,
we will add the worker params into the test params. If we do this,
there will be net name and many worker-specific parameters that
could overwrite test parameters like test name, shortname, and others
by mistake. This will only happen sometimes because if the test is
picked by a worker that doesn't provide any setup, the test name will
not be overwritten. If however a test is picked by a worker that has
also picked some of its setup tests, the worker name will overwrite
the vmnet name when we create the vmnet parameters for the test.

Let's avoid such uncertainty and skip updating the vmnet with any
parameters that will not be used by the vmnet.

While it makes sense for a worker to have a name in its parameter
and for vmnet parameters to be filtered by the worker name, note
that the above is setup providing nets and not the net where the
test will be executed.