Convert murky extra multi-vm restriction to formal vm-variant one
The previously untested leftover functionality of using auxiliary
objects to further restrict non-unique parents while a unique one
is expected is not tested and thus formalized functionality where
the variants of any auxiliary objects will be used to properly
select the correct parent node variant for a dependency object as
long as these participate in the setup test. This supports the
previous restriction but it is not optional based on a vague notion
of "not unique yet" and instead always enforced as long as the
vms or other test objects are confirmed to participate in the setup.

This means that a tutorial_get.explicit_noop test node using CentOS
variant for vm1 will only accept tutorial_gui.client_noop..CentOS
setup for its vm2 object dependency and not tutorial_gui..Fedora
since vm1 participates in the setup and its variant enforces a
restriction to the setup variants involved.